Add Classroom' settings page (#49)
* Added edit classroom detail section

* Decrease content size to 5xl

* Redirect cancel button back to classroom page

* Added subheader to classroom general setting

* Added group section table

* Perform clean install

* Wrapped group slug in CodeLine

* add name,sirname,phone on database

* stasj

* Update bun lockb

* Merge main (squashed):

commit 0a2c228e7937592e97c75211b2d633d75bb65f1b
Author: Peerasin <116646446+Gapraomoosap@users.noreply.github.com>
Date:   Tue Mar 12 15:22:29 2024 +0700

    Register using Email/Password (frontend) (#32)

    * testing

    * create register form

    * change design

    * Add css and picture

    * install bun

    * Add event on submit button

    * connect with api

    * remove console.log

    * change name

    * update

    * add otp component

    * fix lint bug

    * Change otp input

    * Connect otp with api

    * link register with otp verify

    * add valid input alert

    * fix Eslint bug

    * Removed node_modules from root 💀

    * Run ESLint & Prettier

    * Refactor...

    * chage review require

    ---------


commit aa6e751ad4ae11593200f9b3bf40a9fd019b3f2b
Author: tonkaew131 <40863067+tonkaew131@users.noreply.github.com>
Date:   Tue Mar 12 15:17:32 2024 +0700

    Support ticket system (#31)

    * Added ticker database schema

    * Added websocket ticket route api

    * Moved ticket chat to dynamic route

    * Added landing page for ticker support

    * Added support ticket ui

    * Added footer to support page

    * Update ticket link in footer

    * Fixed landing mobile responsive

    * Use daisy ui as chat bubble

    * Added chat history api

    * Added ticket list box

    * Added linked to ticket chat

    * Fixed support should see all tickets

    * Show supporter chat box they're supporter

    * Added createTicket hooks

    * Moved create ticket form into its component

    * Show ticket date

    * Added chat list beside chat box

    * Fixed overflow ticket title

    * Sorted chat history by date

    * Added ability to close ticket

    * Refresh ticket when create ticket

    * Scroll new msg into view

    * Fixed close ticket not update chatbox

    * Create env for websocket url

commit e8fd68e2b9fe62349e508ae356e0e48e64c6c20a
Author: tonkaew131 <40863067+tonkaew131@users.noreply.github.com>
Date:   Fri Mar 8 23:01:58 2024 +0700

    Rename docker file extension to yaml (#34)

    * Added deploy action

    * Rename docker file extension to yaml

* Clean install package

* Clean install package

* stash

* If group slug is empty then create invite code as default group

* Added view my classroom (#44)

* Added get my class api route & its hook

* Fixed initial script cannot run because two tables depend on eather other

* Fixed eslint error

* Changed /classroom to /c in useClassroom hook

* Added classroom card

* Added classroom card footer

* Show teach room

* Added professor name to classroom card

* When create classroom, description isn't required anymore

* Added danger zone to classroom settings

* Added delete classroom button

* Show group list

* Added create group button

* Fixed enter on create group name doesn't update group table

* Added manage group member button & its form

* Added Firstname, Lastname, Phone number to Register API (#45)

* add nane,surname,phone

* update schema

* format phone

* create ui member list

* Fixed Register API, insert on wrong column (name/surname -> first_name/last_name)

* fix name not match

* Refactor code

---------


* Added usehooks-ts

* Added student/teacher member API route

* Added search feature to student/teacher list API route

* Added add new member combobox

* Added add/remove member from group

* Disallow manage group member  for a default group

* Added delete group dialog

* Added delete group button

---------

+++ backend/src/routes/classroom-group.ts
@@ -86,7 +86,8 @@ export const classroomGroupRoute = new Elysia({ prefix: '/c' }).group(
 
                 const classroom = await sql`
                     SELECT
-                        teach.classroom_id
+                        teach.classroom_id,
+                        classroom.default_group
                     FROM teach
                     INNER JOIN classroom
                         ON classroom.id = teach.classroom_id
@@ -105,7 +106,16 @@ export const classroomGroupRoute = new Elysia({ prefix: '/c' }).group(
                     };
                 }
 
-                const { classroom_id: classroomId } = classroom[0];
+                const {
+                    classroom_id: classroomId,
+                    default_group: defaultGroupId,
+                } = classroom[0];
+
+                const [defaultGroup] = await sql`
+                SELECT
+                    slug
+                FROM classroom_group
+                WHERE id = ${defaultGroupId}`;
 
                 const group = await sql`
                 SELECT
@@ -120,6 +130,7 @@ export const classroomGroupRoute = new Elysia({ prefix: '/c' }).group(
                 return {
                     status: 'success',
                     data: group,
+                    default_group: defaultGroup.slug,
                 };
             })
             .group('/:groupSlug', (subapp) =>
@@ -164,8 +175,14 @@ export const classroomGroupRoute = new Elysia({ prefix: '/c' }).group(
 
                         const members = await sql`
                             SELECT
-                                user_id AS "userId"
+                                auth_user.id,
+                                auth_user.first_name AS "firstName",
+                                auth_user.last_name AS "lastName",
+                                auth_user.email,
+                                auth_user.phone_number AS "phoneNumber"
                             FROM classroom_group_member
+                            INNER JOIN auth_user
+                                ON auth_user.id = classroom_group_member.user_id
                             WHERE group_id = ${groupId}
                         `;
 
@@ -190,7 +207,8 @@ export const classroomGroupRoute = new Elysia({ prefix: '/c' }).group(
 
                             const group = await sql`
                             SELECT
-                                classroom_group.id
+                                classroom_group.id,
+                                classroom.default_group
                             FROM classroom_group
                             INNER JOIN classroom
                                 ON classroom.id = classroom_group.classroom_id
@@ -212,8 +230,19 @@ export const classroomGroupRoute = new Elysia({ prefix: '/c' }).group(
                                 };
                             }
 
-                            const { id: groupId } = group[0];
+                            const { id: groupId, default_group: defaultGroup } =
+                                group[0];
                             const { user_id: studentId } = body;
+
+                            if (groupId === defaultGroup) {
+                                set.status = 400;
+
+                                return {
+                                    status: 'error',
+                                    message:
+                                        'You cannot change the member of the default group.',
+                                };
+                            }
 
                             await sql`
                             INSERT INTO classroom_group_member
@@ -248,7 +277,8 @@ export const classroomGroupRoute = new Elysia({ prefix: '/c' }).group(
 
                             const group = await sql`
                             SELECT
-                                classroom_group.id
+                                classroom_group.id,
+                                classroom.default_group
                             FROM classroom_group
                             INNER JOIN classroom
                                 ON classroom.id = classroom_group.classroom_id
@@ -270,8 +300,19 @@ export const classroomGroupRoute = new Elysia({ prefix: '/c' }).group(
                                 };
                             }
 
-                            const { id: groupId } = group[0];
+                            const { id: groupId, default_group: defaultGroup } =
+                                group[0];
                             const { user_id: studentId } = body;
+
+                            if (groupId === defaultGroup) {
+                                set.status = 400;
+
+                                return {
+                                    status: 'error',
+                                    message:
+                                        'You cannot change the member of the default group.',
+                                };
+                            }
 
                             await sql`
                             DELETE FROM
